Task2_M2KR/requirements.txt
accelerate==1.4.0
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
colpali_engine==0.3.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.3.2
debugpy==1.8.12
decorator==5.2.1
dill==0.3.8
exceptiongroup==1.2.2
executing==2.1.0
faiss-gpu
fastembed==0.6.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
GPUtil==1.4.0
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
libarchive-c==5.1
loguru==0.7.3
markdown-it-py==3.0.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.28.0
nest_asyncio==1.6.0
numpy==2.2.3
onnxruntime==1.20.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.14.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pip==25.0.1
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py_rust_stemmers==0.1.5
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
qdrant-client==1.13.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
safetensors==0.5.3
scipy==1.15.2
setuptools==75.8.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack_data==0.6.3
tenacity==9.0.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
wheel==0.45.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0